extract `position_inner.rs` from `position.rs`
