+++ src/tests/submit_market_buy_order.rs
@@ -1,8 +1,8 @@
 use hashbrown::HashMap;
 
 use crate::{
-    mock_exchange::MockTransactionAccounting, mock_exchange_linear, position::PositionInner,
-    prelude::*, risk_engine::RiskError,
+    mock_exchange::MockTransactionAccounting, mock_exchange_linear, prelude::*,
+    risk_engine::RiskError,
 };
 
 #[test]
